rgw_a: build reorganization
Reorganize some dependencies shared by the librgw.so and radosgw
main units.  The librgw.so build no longer builds a main(...)
unit.

Build time should be slightly reduced, as several shared dependencies
have been moved into rgw_a.

There are some other cleanups, e.g., rgw_signal.{h,cc}.
